Divided main code to files
