+++ src/phonebook/Main.java
@@ -138,140 +138,3 @@ public class Main {
         return "Searching time: " + formattedTime(millis);
     }
 }
-
-
-enum Algorithm {
-    LINEAR, JUMP, BINARY
-}
-
-abstract class Search {
-    public static int listSearch(ArrayList<String> directory, ArrayList<String> find, Algorithm algorithm) {
-        int found = 0;
-
-        for (String target : find) {
-            switch (algorithm) {
-                case LINEAR:
-                    found += linearSearch(directory, target) ? 1 : 0;
-                    break;
-                case JUMP:
-                    found += jumpSearch(directory, target) ? 1 : 0;
-                    break;
-                case BINARY:
-                    found += binarySearch(directory, target) ? 1 : 0;
-                    break;
-                default:
-                    return 0;
-            }
-        }
-
-        return found;
-    }
-
-    public static boolean linearSearch(ArrayList<String> directory, String find) {
-        for (String contact : directory) {
-            if (contact.equalsIgnoreCase(find)) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    public static boolean jumpSearch(ArrayList<String> directory, String find) {
-        int blockSize = (int) Math.floor(Math.sqrt(directory.size()));
-
-        int index = 0;
-        while (index < directory.size()) {
-            if (directory.get(index).equalsIgnoreCase(find)) return true;
-
-            if (directory.get(index).compareToIgnoreCase(find) > 0) {
-                for (int i = index - 1; i > index - blockSize; i--) {
-                    if (directory.get(index).equalsIgnoreCase(find)) return true;
-                }
-
-                return false;
-            }
-
-            index += blockSize;
-        }
-
-        for (int i = directory.size() - 1; i > directory.size() - blockSize; i--) {
-            if (directory.get(i).equals(find)) return true;
-        }
-
-        return false;
-    }
-
-    public static boolean binarySearch(ArrayList<String> directory, String find) {
-        int left = 0;
-        int right = directory.size() - 1;
-
-        while (left <= right) {
-            int mid = (left + right) / 2;
-            String midElement = directory.get(mid);
-
-            if (midElement.equalsIgnoreCase(find)) return true;
-            else if (midElement.compareToIgnoreCase(find) > 0) right = mid - 1;
-            else left = mid + 1;
-        }
-
-        return false;
-    }
-}
-
-abstract class Sort {
-    public static boolean bubbleSort(ArrayList<String> list, long maxTime) {
-        long millis = System.currentTimeMillis();
-        long limit = maxTime * 10;
-
-        for (int size = list.size(); size > 1; size--) {
-            if (System.currentTimeMillis() - millis > limit) {
-                return false;
-            }
-
-            for (int i = 0; i < size; i++) {
-                if (System.currentTimeMillis() - millis > limit) {
-                    return false;
-                }
-
-                if ((i + 1 < size) && (list.get(i).compareToIgnoreCase(list.get(i + 1)) == 1)) {
-                    String tmp = list.get(i);
-
-                    list.set(i, list.get(i + 1));
-                    list.set(i + 1, tmp);
-                }
-            }
-        }
-
-        return true;
-    }
-
-    public static void quickSort(ArrayList<String> list, int left, int right) {
-        if (left < right) {
-            String pivot = list.get(right);
-
-            int pointer = left - 1;
-
-            for (int i = left; i <= right - 1; i++) {
-                if (list.get(i).compareToIgnoreCase(pivot) < 0) {
-                    pointer++;
-
-                    swap(list, pointer, i);
-                }
-            }
-
-            swap(list, pointer + 1, right);
-
-            int pi = (pointer + 1);
-
-            quickSort(list, left, pi - 1);
-            quickSort(list, pi + 1, right);
-        }
-    }
-
-    private static void swap(ArrayList<String> list, int i, int j) {
-        String temp = list.get(i);
-        list.set(i, list.get(j));
-        list.set(j, temp);
-    }
-}
